+++ src/services/editor/editor-icons.ts
